Working now with the Tradfri gateway
This communication requires a compilation of WolfSSL with the WOLFSSL_STATIC_PSK option enabled.

We use an extension of the library project to include the compilation switch (this impedes using the library as Alire dependency, but since it's included as a Git submodule, there's nothing to lose).

For example, this command will turn one of the lightbulbs to on (once having the correct identity and PSK in the code):
```
bin/coap_client -m put -e '{"3311": [{"5850": 1}]}' "coaps://tradfri-gw:5684/15001/65559"
```