Fix assertion error with option values of 13 or 14 bytes and option delta lower than 13
The reason was a hole in the RFLX spec for the option format. The conditions for going to Option_Length_Extended{8, 16} from after Option_Length weren't defined.

This was being detected by SPARK in silver mode when setting Option_Length, with a failing precondition with `Field_Condition`. Now, it doesn't complain.

Now, there are 19 failed VC in `rflx-coap_client-session.adb`.