Allow building with assertions enabled
We need to repeat the same `Assertion_Policy` pragmas  that are included in `SPARK.Containers.Formal.Unbounded_Vectors` before our instantiation of `Generic_Sorting`, otherwise the contracts in that package are executed when building in the validation profile for all dependencies.

This means, that building the client with `alr build --profiles="*=validation"` no longer breaks the tests.

This is a workaround for "CS0040544 #4830 Constraint error in SPARKLib and linking problem", but only for the contrainer error in SPARKLib. The linker problem regarding `Check_Or_Fail` is still present.