Fix server test for non-Linux systems
It was valid to open this port on Windows and macOS.

Change to a test about the validity of the port number itself. Detected a problem with the precondition for the non-SPARK function used for converting from string to number. The 'Width is always one more than the actual width for positive numbers (character for the minus sign).